# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Client Wealth Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  client: 'client',
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
  clientId?: number | null;
  name: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface CreateClientBody {
  name: string;
  email?: string;
  phone?: string;
  username: string;
  password: string;
}

export interface UpdateClientBody {
  name?: string;
  email?: string;
  phone?: string;
}

export interface Client {
  id: number;
  name: string;
  email?: string | null;
  phone?: string | null;
  username: string;
  createdAt: string;
}

export type AssetType = typeof AssetType[keyof typeof AssetType];


export const AssetType = {
  mutual_fund: 'mutual_fund',
  stock: 'stock',
  fixed_deposit: 'fixed_deposit',
  recurring_deposit: 'recurring_deposit',
  provident_fund: 'provident_fund',
  cash_bank: 'cash_bank',
} as const;

export type CreateAssetBodyData = { [key: string]: unknown };

export interface CreateAssetBody {
  assetType: AssetType;
  data: CreateAssetBodyData;
  familyMemberId?: number | null;
}

export type UpdateAssetBodyData = { [key: string]: unknown };

export interface UpdateAssetBody {
  data: UpdateAssetBodyData;
  familyMemberId?: number | null;
}

export type AssetData = { [key: string]: unknown };

export interface Asset {
  id: number;
  clientId: number;
  familyMemberId?: number | null;
  assetType: AssetType;
  data: AssetData;
  value: number;
  createdAt: string;
}

export type LoanType = typeof LoanType[keyof typeof LoanType];


export const LoanType = {
  home_loan: 'home_loan',
  car_loan: 'car_loan',
  personal_loan: 'personal_loan',
  education_loan: 'education_loan',
  business_loan: 'business_loan',
  other: 'other',
} as const;

export interface CreateLiabilityBody {
  loanType: LoanType;
  lenderName: string;
  totalLoanAmount: number;
  outstandingAmount: number;
  interestRate: number;
  emi: number;
  startDate?: string;
  endDate?: string;
  familyMemberId?: number | null;
  notes?: string | null;
}

export interface UpdateLiabilityBody {
  loanType?: LoanType;
  lenderName?: string;
  totalLoanAmount?: number;
  outstandingAmount?: number;
  interestRate?: number;
  emi?: number;
  startDate?: string;
  endDate?: string;
  familyMemberId?: number | null;
  notes?: string | null;
}

export interface Liability {
  id: number;
  clientId: number;
  familyMemberId?: number | null;
  loanType: LoanType;
  lenderName: string;
  totalLoanAmount: number;
  outstandingAmount: number;
  interestRate: number;
  emi: number;
  startDate?: string | null;
  endDate?: string | null;
  /** Additional details or specific category breakdown */
  notes?: string | null;
  createdAt: string;
}

export interface AssetBreakdownItem {
  assetType: AssetType;
  total: number;
  count: number;
}

export interface ClientSummary {
  clientId: number;
  clientName: string;
  totalAssets: number;
  totalLiabilities: number;
  netWorth: number;
  assetBreakdown: AssetBreakdownItem[];
}

export interface DashboardOverview {
  totalClients: number;
  totalAUM: number;
  totalLiabilities: number;
  totalNetWorth: number;
  assetBreakdown: AssetBreakdownItem[];
  recentClients: Client[];
}

export type FamilyMemberRelation = typeof FamilyMemberRelation[keyof typeof FamilyMemberRelation];


export const FamilyMemberRelation = {
  Parent: 'Parent',
  Spouse: 'Spouse',
  Child: 'Child',
} as const;

export interface FamilyMember {
  id: number;
  clientId: number;
  name: string;
  dob?: string | null;
  phone?: string | null;
  relation: FamilyMemberRelation;
  createdAt: string;
}

export type CreateFamilyMemberBodyRelation = typeof CreateFamilyMemberBodyRelation[keyof typeof CreateFamilyMemberBodyRelation];


export const CreateFamilyMemberBodyRelation = {
  Parent: 'Parent',
  Spouse: 'Spouse',
  Child: 'Child',
} as const;

export interface CreateFamilyMemberBody {
  name: string;
  dob?: string;
  phone?: string;
  relation: CreateFamilyMemberBodyRelation;
}

export type UpdateFamilyMemberBodyRelation = typeof UpdateFamilyMemberBodyRelation[keyof typeof UpdateFamilyMemberBodyRelation];


export const UpdateFamilyMemberBodyRelation = {
  Parent: 'Parent',
  Spouse: 'Spouse',
  Child: 'Child',
} as const;

export interface UpdateFamilyMemberBody {
  name?: string;
  dob?: string;
  phone?: string;
  relation?: UpdateFamilyMemberBodyRelation;
}