# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Client Wealth Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /clients/{clientId}/assets:
    get:
      operationId: listClientAssets
      tags: [assets]
      summary: List all assets for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createClientAsset
      tags: [assets]
      summary: Add an asset for a client (admin only)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetBody"
      responses:
        "201":
          description: Asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /clients/{clientId}/assets/{assetId}:
    put:
      operationId: updateClientAsset
      tags: [assets]
      summary: Update an asset (admin only)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssetBody"
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteClientAsset
      tags: [assets]
      summary: Delete an asset (admin only)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, client]
        clientId:
          type: integer
          nullable: true
        name:
          type: string
      required:
        - id
        - username
        - role
        - name
...
    AssetType:
      type: string
      enum:
        - mutual_fund
        - stock
        - fixed_deposit
        - recurring_deposit
        - provident_fund
        - cash_bank
    CreateAssetBody:
      type: object
      properties:
        assetType:
          $ref: "#/components/schemas/AssetType"
        data:
          type: object
          additionalProperties: true
        familyMemberId:
          type: integer
          nullable: true
      required:
        - assetType
        - data
    UpdateAssetBody:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
        familyMemberId:
          type: integer
          nullable: true
      required:
        - data
    Asset:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        familyMemberId:
          type: integer
          nullable: true
        assetType:
          $ref: "#/components/schemas/AssetType"
        data:
          type: object
          additionalProperties: true
        value:
          type: number
        createdAt:
          type: string
      required:
        - id
        - clientId
        - assetType
        - data
        - value
        - createdAt
...